api/Hub.Shell.Api/Controllers/PingController.cs
using System.Reflection;
using Hub.Shell.Contracts;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;

namespace Hub.Shell.Api.Controllers
{
    [ApiController, AllowAnonymous]
    [Route("[controller]")]
    public class PingController : ControllerBase
    {
        private readonly ILogger<PingController> _logger;

        public PingController(ILogger<PingController> logger)
        {
            _logger = logger;
        }

        [HttpGet]
        public IActionResult Ping()
        {
            var response = new PingContract
            {
                ApplicationVersion = Assembly.GetExecutingAssembly()
                    .GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion
            };

            return new OkObjectResult(response);
        }
    }
}